refactor(factory-ai-droid): isolate repeated identity state
